pyproject.toml
[project]
name = "cs520-exercise3-llm-specification-generation"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "dotenv>=0.9.9",
    "openai>=2.8.1",
    "psutil>=7.1.3",
    "pytest>=9.0.1",
    "pytest-cov>=7.0.0",
    "pyyaml>=6.0.3",
]
